Move filters into separate directory
