+++ cg/store/api/findbusinessdata.py
@@ -23,9 +23,9 @@ from cg.store.models import (
     Pool,
     Sample,
 )
-from cg.store.status_flow_cell_filters import apply_flow_cell_filter, FlowCellFilters
-from cg.store.status_case_sample_filters import apply_case_sample_filter
-from cg.store.status_sample_filters import apply_sample_filter
+from cg.store.filters.status_flow_cell_filters import apply_flow_cell_filter, FlowCellFilters
+from cg.store.filters.status_case_sample_filters import apply_case_sample_filter
+from cg.store.filters.status_sample_filters import apply_sample_filter
 
 LOG = logging.getLogger(__name__)
