+++ tests/store/test_status_collaboration_filters.py
@@ -1,6 +1,6 @@
 from cg.store import Store
 from cg.store.models import Collaboration
-from cg.store.status_collaboration_filters import filter_collaboration_by_internal_id
+from cg.store.filters.status_collaboration_filters import filter_collaboration_by_internal_id
 
 
 def test_get_collaboration_by_internal_id(base_store: Store, collaboration_id: str):
